+++ CHANGES.md
@@ -6,6 +6,12 @@ Version 0.13.2
 
 To be released.
 
+ - When a reorg happens, `Swarm<T>` now broadcasts a reorged chain tip first
+   before rendering. [[#1385], [#1415]]
+
+[#1385]: https://github.com/planetarium/libplanet/issues/1385
+[#1415]: https://github.com/planetarium/libplanet/pull/1415
+
 
 Version 0.13.1
 --------------